refactor: 경로 탐색 역할 Lines -> Path로 변경, 요금 계산 책임 분리
